Fixes #44
This is a bit hacky and basically makes two passes through the SADFace
nodes. On the first pass we process atoms and schemes (primary nodes)
and on the second pass we process named args (secondary nodes). A
better solution would cache named args on the first pass and then only
process that cache of nodes in a second pass. This would greatly save
processing effort on larger argument graphs.